jordan: added rotational velocity during collision
BIG UPDATE!

Now the point of contact actually affects rotational velocity, so objects will
rotate after being collided and looks much more realistic.

Point of contact currently is computed by checking the intersection of every
edge in object 1 with every surface/polygon on object 2, and vice versa, then
averaged to appxoimate the 'average force.' This is alright, but can result in
non-ideal situations when the object is about to come to rest and starts
spinning on its own, as rotational velocity is not yet local but instead global,
which is not really true for point of force applied.

Friction coefficient is also weird. Simulation becomes a lot less stable when
the coefficient is changed from 0.03 to 0.05 with 20 objects and a 30x20x30
room.

A lot of manually tweaking the constants has been done to get something that
'looks right.' Though the laws are still based on classical mechanics, the
constants and weighing no longer do. I have to artificially lower the influence
of rotational impulse, as other wise objects stop bouncing and just rotate,
which honestly makes a lot of sense for boxes in real life, but not so great
when everything is modeled as boxes.

Added some extra toggles for initializing the scene (now random with some
specified number of objects) and toggling the BLENDER. BRRRRRRRRRRRRRRRRRRR

I do want to eventually implement local gravitational pulls, i.e., a tilted box
will eventually drop to its equilibrium position because the parts of it not in
direct contact with the floor still has gravity pulling on it (in terms of net
force). There was an attempt made for this commit, but I ultimately removed it
as it was way too buggy.

^ For record's sake, I simply tried adding a rotation acceleration that pushed
the boxes' rotation to some equilibrium position (axis-aligned), where the
acceleration is proportional to how 'off' the box is to said equilibrium. Turns
out, that is EXACTLY the condition for simple harmonic motion and in fact can
just make the boxes waddle back and forth or, worse, start spinning.

There are a lot more changes, but I'm tired and want to sleep, so I will do
exactly that.

Jordan Lin, 2022/5/19 23:47

P.S. Oh crap I have not looked at tomorrow's CS 33 LA worksheet yet. Need to do
that soon D: