jordan: add composite object to physics
Now the physics engine supports multiple objects that are treated as the same
object with still individual hitboxes, so composite objects (objects made from
smaller objects) can now be done, and shapes like tables can actually have
accurate hitboxes.

Now, this.objects no longer stores an array of objects but instead an array of
object groups (composite objects), where each object group or composite object
is itself an array of objects. Hence, to initialize just one object, one now has
to nest it inside an array.

New attribute for object has also been added, that is, rotation_model. Because
the point at which an objec group rotates---the center of mass---is different
than the center of the individual object models, rotation_model is added for the
rotation of the model itself, and rotation now controls the rotation for the
entire object group. Note that the engine expects all movement properties, e.g.,
rotation, rotation_velocity, velocity, acceleration, etc., of the objects inside
an object group to be the same as all other objects inside the same object
group, and if an object group were to be treated like a wall, all objects inside
the object group must be a wall as well.